+++ components/Document/DocumentEditor.js
@@ -1,13 +1,13 @@
 import { inject, observer } from 'mobx-react'
 import React, { Component, useState, useEffect, useRef } from 'react'
 import { withForm } from './DocumentForm'
-import EditorView from './Editor/views/EditorView'
+import EditorView from '../Editor/views/EditorView'
 import { convertFromRaw, EditorState } from 'draft-js'
 import { toJS } from 'mobx'
 import { Button, ButtonGroup } from '@material-ui/core'
 import CircularProgress from "@material-ui/core/CircularProgress";
-import { RichEditor } from './RichEditor'
-import { NullReferenceError } from './Errors'
+import { RichEditor } from '../RichEditor'
+import { NullReferenceError } from '../Errors'
 import { compose } from 'recompose'
 
 // Document editor is a shim that connects our feature-rich DraftJS editor to
