+++ components/Document/FileBrowser.js
@@ -1,0 +1,9 @@
+/** Basic File Browser for uploads and conversions */
+const FileBrowser = ({ onSelected }) => {
+    const setFile = (event) => {
+        const file = event.target.files[0];
+        onSelected(file);
+    }
+    return <input type="file" onChange={setFile} />;
+}
+export default FileBrowser;
