+++ components/Document/index.js
@@ -1,0 +1,9 @@
+import DocumentDetails from './DocumentDetails'
+import DocumentEditor from './DocumentEditor'
+import DocumentForm from './DocumentForm'
+
+export {
+    DocumentDetails,
+    DocumentForm,
+    DocumentEditor
+}
