+++ components/Editor/functions/WrittenFile.js
