+++ components/Errors/index.js
@@ -11,4 +11,5 @@ const ArgumentNullReferenceMessage = argName => `${argName || "Argument"} cannot
 export class EditorStateNotFoundError extends ExtendableError { constructor() { super(EditorStateNotFoundMessage) } }
 export class NullReferenceError extends ExtendableError { constructor() { super(NullReferenceMessage) } }
 export class ArgumentNullReferenceError extends ExtendableError { constructor(argname = null) { super(ArgumentNullReferenceMessage(argname)) } }
+export default ExtendableError
