+++ components/RichEditor/buttons/PrimaryButton.js
@@ -1,7 +1,11 @@
-import { Button } from   '@material-ui/core'
-export const PrimaryButton = props => {
+import { Button } from '@material-ui/core'
 
-  // const palette = props.palette;
+
+
+
+const PrimaryButton = props => {
+
+  const palette = props.palette;
 
   // if(!palette)
   //   throw new Error("Missing a color palette!");
@@ -22,11 +26,10 @@ export const PrimaryButton = props => {
     className += " RichEditor-activeButton";
   }
 
-
   return (
     <span className={className} onMouseDown={onToggle}>
       <button onClick={props.onClick}>{props.children || props.label}</button>
-      {/* <style jsx>{`
+      {pallete && <style jsx>{`
         button{
             border: "2px solid" + ${pallete.primary.teal};
             color:${pallete.primary.blue};
@@ -52,14 +55,17 @@ export const PrimaryButton = props => {
             margin:0.2em auto;
            }
           } 
-      `}</style> */}
+      `}</style>}
     </span>
   );
+
+
+
 
   // return (
   //   <Button onClick={props.onClick}>{props.children || props.label}</Button>
   //   );
-    {/* <span className={className} onMouseDown={onToggle}>
+  {/* <span className={className} onMouseDown={onToggle}>
       <button onClick={props.onClick}>{props.children || props.label}</button>
       <style jsx>{
         
@@ -117,3 +123,7 @@ export const PrimaryButton = props => {
       `}</style>
     </span> */}
 };
+
+
+export default PrimaryButton;
+
