+++ components/RichEditor/buttons/index.js
@@ -1,0 +1,7 @@
+import PrimaryButton  from './PrimaryButton';
+import SubmitButton from './SubmitButton'
+
+export {
+    PrimaryButton,
+    SubmitButton
+}
