+++ package.json
@@ -63,14 +63,15 @@
     "react-toastify": "^5.5.0",
     "recompose": "^0.30.0",
     "snake-case": "^3.0.3",
-    "validatorjs": "^3.18.1"
+    "validatorjs": "^3.18.1",
+    "wpapi": "^1.2.1"
   },
   "resolutions": {
     "jss": "10.0.0"
   },
   "devDependencies": {
     "@babel/plugin-proposal-class-properties": "^7.7.4",
-    "netlify-cli": "^2.47.0",
-    "@babel/plugin-proposal-decorators": "^7.7.4"
+    "@babel/plugin-proposal-decorators": "^7.7.4",
+    "netlify-cli": "^2.47.0"
   }
 }
