+++ services/firebase/index.js
@@ -9,6 +9,15 @@ import { observable, toJS } from 'mobx'
 
 import { convertFile } from '../../components/Editor/functions/converter'
 
+// const config = {
+//     apiKey: process.env.REACT_APP_API_KEY,
+//     authDomain: process.env.REACT_APP_AUTH_DOMAIN,
+//     databaseURL: process.env.REACT_APP_DATABASE_URL,
+//     projectId: process.env.REACT_APP_PROJECT_ID,
+//     storageBucket: process.env.REACT_APP_STORAGE_BUCKET,
+//     messagingSenderId: process.env.REACT_APP_MESSAGING_SENDER_ID,
+// };
+
 class Firebase {
 
     @observable authUser = null
@@ -154,7 +163,15 @@ class Firebase {
     }
 
     checkout = async id => {
-        console.log('chekcout document', id)
+        // console.log('checking out document', id)
+        // const document = new Document(`sessions/${id}`)
+        // await document.fetch()
+        // const { status } = document.data
+
+        // if (['not-started', 'in-progress'].includes(status)) {
+        //     this.redirect("/scribe/edit/[doc]", `/scribe/edit/${id}`)
+        // }
+
         return new Promise(async (resolve, reject) => {
             console.log(`Checkout Document by ID: ${id}`)
             let requestedDocument = new Document(`sessions/${id}`)
